fix(billing): refactor payment list components
