+++ apps/deploy-web/src/components/shared/PaymentMethodCard/index.ts
@@ -1,0 +1,1 @@
+export { PaymentMethodCard } from "./PaymentMethodCard";
