+++ apps/deploy-web/src/components/shared/PaymentMethodsList/index.ts
@@ -1,0 +1,1 @@
+export { PaymentMethodsList } from "./PaymentMethodsList";
