+++ apps/deploy-web/src/components/user/payment/index.ts
@@ -1,4 +1,3 @@
 export * from "./AddPaymentMethodPopup";
 export * from "./DeletePaymentMethodPopup";
 export * from "./PaymentForm";
-export * from "./PaymentMethodsList";
